replace old create message code and move jobs url
