+++ apps/messaging-api/services/users/import/import-users.ts
@@ -11,7 +11,7 @@ import { parseFile, writeToBuffer } from "fast-csv";
 import { Pool, PoolClient } from "pg";
 import { isNativeError } from "util/types";
 import { mapUsers } from "./map-users";
-import { sendInvitationsForUsersImport } from "./send-invitations";
+import { sendInvitationsForUsersImport } from "../invitations/send-invitations";
 import { PostgresDb } from "@fastify/postgres";
 import { BadRequestError, ServerError } from "shared-errors";
